Add files via upload
This is the arduino software for the TTGO unit.  This is an updated version of the software.  Version 1.4. It adds the following features:

In the CIV_template.ino file there is an option to set HOME to 1 or 0.  I use this so I can have two TTGO units.  Setting HOME = 1 is the normal function of the TTGO at the QTH.  Setting HOME = 0 is for the TTGO that I take in the field.  This utilized a different table for TX Power in the z_userprog.info file that I use lower power to save battery.  It also doesn't automatically change the RF Gain  as it will be different when doing remote operations

In the z_userprog.info there is an option to set frequencies for using a HOTSPOT and having it set the TX Power for those frequencies.